Progress on integrating Stripe
As of now, I have not yet figure out how to test this rigorously (I haven't been able to get Capybara to select the Pay By Card button. I have tried selecting by text and CSS class to no avail as of yet). Next step is to create an order in the database when the user checks out.